refactor sql into, #1895
