+++ packages/core/src/abap/2_statements/expressions/index.ts
@@ -147,7 +147,7 @@ export * from "./sql_order_by";
 export * from "./sql_path";
 export * from "./sql_source_simple";
 export * from "./sql_source";
-export * from "./sql_target_table";
+export * from "./sql_into_table";
 export * from "./sql_target";
 export * from "./string_template_formatting";
 export * from "./string_template";
