+++ packages/core/src/abap/2_statements/expressions/select.ts
@@ -1,21 +1,13 @@
-import {seq, per, opt, alt, tok, str, starPrio, Expression, altPrio, optPrio, ver} from "../combi";
-import {WParenLeftW, WParenLeft} from "../../1_lexer/tokens";
-import {SQLTarget, SQLFieldList, SQLFrom, SQLCond, SQLSource, DatabaseConnection, SQLTargetTable, SQLOrderBy, SQLHaving, SQLForAllEntries} from ".";
+import {seq, per, opt, alt, str, Expression, altPrio, optPrio, ver} from "../combi";
+import {SQLFieldList, SQLFrom, SQLCond, SQLSource, DatabaseConnection, SQLIntoTable, SQLOrderBy, SQLHaving, SQLForAllEntries} from ".";
 import {Version} from "../../../version";
 import {IStatementRunnable} from "../statement_runnable";
 import {SQLGroupBy} from "./sql_group_by";
+import {SQLIntoStructure} from "./sql_into_structure";
 
 export class Select extends Expression {
   public getRunnable(): IStatementRunnable {
-
-    const intoList = seq(alt(tok(WParenLeft), tok(WParenLeftW)),
-                         starPrio(seq(SQLTarget, ",")),
-                         SQLTarget,
-                         ")");
-    const intoSimple = seq(optPrio("CORRESPONDING FIELDS OF"),
-                           SQLTarget);
-
-    const into = alt(seq("INTO", altPrio(intoList, intoSimple)), SQLTargetTable);
+    const into = alt(SQLIntoStructure, SQLIntoTable);
 
     const where = seq("WHERE", SQLCond);
