+++ packages/core/src/abap/2_statements/expressions/select_loop.ts
@@ -1,26 +1,16 @@
-import {seq, per, opt, alt, tok, ver, star, Expression, optPrio} from "../combi";
-import {WParenLeftW, WParenLeft} from "../../1_lexer/tokens";
-import {SQLSource, SQLFrom, DatabaseTable, Dynamic, SQLCond, SQLFieldName, SQLAggregation, SQLTargetTable, SQLGroupBy, SQLForAllEntries} from ".";
+import {seq, per, opt, alt, ver, star, Expression, optPrio} from "../combi";
+import {SQLSource, SQLFrom, DatabaseTable, Dynamic, SQLCond, SQLFieldName, SQLAggregation, SQLIntoTable, SQLGroupBy, SQLForAllEntries} from ".";
 import {Version} from "../../../version";
 import {IStatementRunnable} from "../statement_runnable";
 import {SQLOrderBy} from "./sql_order_by";
 import {SQLHaving} from "./sql_having";
-import {SQLTarget} from "./sql_target";
 import {SQLPath} from "./sql_path";
 import {SQLAsName} from "./sql_as_name";
 import {SQLCase} from "./sql_case";
+import {SQLIntoStructure} from "./sql_into_structure";
 
 export class SelectLoop extends Expression {
   public getRunnable(): IStatementRunnable {
-
-    const intoList = seq(alt(tok(WParenLeft), tok(WParenLeftW)),
-                         star(seq(SQLTarget, ",")),
-                         SQLTarget,
-                         ")");
-    const intoSimple = seq(opt("CORRESPONDING FIELDS OF"), SQLTarget);
-
-    const into = seq("INTO", alt(intoList, intoSimple));
-
     const where = seq("WHERE", SQLCond);
 
     const comma = opt(ver(Version.v740sp05, ","));
@@ -40,7 +30,7 @@ export class SelectLoop extends Expression {
 
     const from2 = seq("FROM", DatabaseTable);
 
-    const tab = seq(SQLTargetTable, alt(pack, seq(from2, pack), seq(pack, from2)));
+    const tab = seq(SQLIntoTable, alt(pack, seq(from2, pack), seq(pack, from2)));
 
     const perm = per(SQLFrom,
                      where,
@@ -51,7 +41,7 @@ export class SelectLoop extends Expression {
                      bypass,
                      SQLGroupBy,
                      SQLForAllEntries,
-                     alt(tab, into));
+                     alt(tab, SQLIntoStructure));
 
     const ret = seq("SELECT",
                     optPrio("DISTINCT"),
