+++ packages/core/src/abap/2_statements/expressions/sql_into_table.ts
@@ -1,0 +1,13 @@
+import {alt, seq, Expression, opt} from "../combi";
+import {SQLTarget} from ".";
+import {IStatementRunnable} from "../statement_runnable";
+
+export class SQLIntoTable extends Expression {
+  public getRunnable(): IStatementRunnable {
+    const into = seq(alt("INTO", "APPENDING"),
+                     opt("CORRESPONDING FIELDS OF"),
+                     "TABLE",
+                     SQLTarget);
+    return into;
+  }
+}
