--- packages/core/src/abap/2_statements/expressions/sql_target_table.ts
@@ -1,14 +1,0 @@
-import {alt, seq, Expression, opt} from "../combi";
-import {SQLTarget} from ".";
-import {IStatementRunnable} from "../statement_runnable";
-
-export class SQLTargetTable extends Expression {
-  public getRunnable(): IStatementRunnable {
-    const intoTable = seq(alt("INTO", "APPENDING"),
-                          opt("CORRESPONDING FIELDS OF"),
-                          "TABLE",
-                          SQLTarget);
-
-    return intoTable;
-  }
-}
