+++ packages/core/src/abap/2_statements/statements/fetch_next_cursor.ts
@@ -1,21 +1,18 @@
 import {IStatement} from "./_statement";
 import {verNot, seq, alt, opt} from "../combi";
-import {SQLTarget, SQLSource, SQLTargetTable} from "../expressions";
+import {SQLSource, SQLIntoTable} from "../expressions";
 import {Version} from "../../../version";
 import {IStatementRunnable} from "../statement_runnable";
+import {SQLIntoStructure} from "../expressions/sql_into_structure";
 
 export class FetchNextCursor implements IStatement {
 
   public getMatcher(): IStatementRunnable {
     const size = seq("PACKAGE SIZE", SQLSource);
 
-    const record = seq("INTO",
-                       opt("CORRESPONDING FIELDS OF"),
-                       SQLTarget);
-
     const ret = seq("FETCH NEXT CURSOR",
                     SQLSource,
-                    alt(record, SQLTargetTable),
+                    alt(SQLIntoStructure, SQLIntoTable),
                     opt(size));
 
     return verNot(Version.Cloud, ret);
