+++ packages/core/test/abap/syntax/syntax.ts
@@ -3873,6 +3873,14 @@ ENDFUNCTION.`;
     expect(issues.length).to.equals(0);
   });
 
+  it.skip("SELECT without INTO, implicit workarea", () => {
+    const abap = `
+    SELECT * FROM foobar WHERE field = 'bar'.
+    ENDSELECT.`;
+    const issues = runProgram(abap);
+    expect(issues.length).to.equals(1);
+  });
+
 // todo, static method cannot access instance attributes
 // todo, can a private method access protected attributes?
 // todo, readonly fields(constants + enums + attributes flagged read-only)
